refactor: detail dialog 컴포넌트 이동
