+++ frontend/src/pages/reservation/search/components/detail-dialog/detail-dialog.tsx
