+++ frontend/src/pages/reservation/search/components/detail-dialog/index.ts
