+++ frontend/src/pages/reservation/search/components/detail-dialog/types.ts
